backend/requirements.txt
uvicorn
fastapi
pymongo
python-jwt
aiosmtplib
passlib
razorpay
jose
httpx
PyJWT